refactor: split app actions into focused modules
